added confirmation when trying to delete a post

+++ admin/includes/view_all_posts.php
@@ -108,7 +108,7 @@
           echo "<td>{$post_date}</td>";
           echo "<td><a href='../post.php?p_id={$post_id}' title='View Post'>View Post</a></td>"; 
           echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}' title='Edit Post'>Edit</a></td>";    
-          echo "<td><a href='posts.php?delete={$post_id}' title='Delete Post'>Delete</a></td>";      
+          echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete'); \" href='posts.php?delete={$post_id}' title='Delete Post'>Delete</a></td>";      
           echo "</tr>";                          }    
         ?>                    
     </tbody>
